Refactor: separate files for trace parsers. Removed unnecessary includes
