+++ include/common_types.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <set>
 #include <unordered_set>
+#include <stdint.h>
 
 typedef int16_t ThreadIdT;
 typedef int64_t EventIdT;
@@ -10,5 +10,4 @@ typedef int32_t SrcLocT;
 typedef int64_t BinEvT;
 
 typedef std::unordered_set<ResourceIdT> ULocksetT;
-
-typedef std::unordered_set<ResourceIdT> UThreadSetT;
+typedef std::unordered_set<ThreadIdT> UThreadSetT;
